+++ axelor-crm/src/main/java/com/axelor/apps/crm/web/EventController.java
@@ -19,8 +19,10 @@ package com.axelor.apps.crm.web;
 
 import com.axelor.apps.base.service.MapService;
 import com.axelor.apps.crm.db.Event;
+import com.axelor.apps.crm.db.EventReminder;
 import com.axelor.apps.crm.db.Lead;
 import com.axelor.apps.crm.db.RecurrenceConfiguration;
+import com.axelor.apps.crm.db.repo.EventReminderRepository;
 import com.axelor.apps.crm.db.repo.EventRepository;
 import com.axelor.apps.crm.db.repo.LeadRepository;
 import com.axelor.apps.crm.db.repo.RecurrenceConfigurationRepository;
@@ -505,6 +507,20 @@ public class EventController {
       if (emailAddress != null) {
         response.setValue("attendees", iCalendarEventService.addEmailGuest(emailAddress, event));
       }
+    } catch (Exception e) {
+      TraceBackService.trace(response, e);
+    }
+  }
+
+  @Transactional(rollbackOn = {Exception.class})
+  public void deleteReminder(ActionRequest request, ActionResponse response) {
+    try {
+      EventReminderRepository eventReminderRepository = Beans.get(EventReminderRepository.class);
+
+      EventReminder eventReminder =
+          eventReminderRepository.find((long) request.getContext().get("id"));
+      eventReminderRepository.remove(eventReminder);
+      response.setCanClose(true);
     } catch (Exception e) {
       TraceBackService.trace(response, e);
     }